requirements.txt
aiofiles==22.1.0
aioredis==2.0.1
aniso8601==9.0.1
anyio==3.7.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.2
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.8
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cache2==0.1.9
fastapi-cli==0.0.7
flake8==5.0.4
frozendict==2.3.4
graphene==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.5.0
httpx==0.28.1
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.6
lxml==4.9.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mccabe==0.7.0
mdurl==0.1.2
obasparql==5.0.2
pendulum==2.1.2
ply==3.11
promise==2.3
pyaml==21.10.1
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.19.1
PyLD==2.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.20
pythonql3==0.9.83
pytzdata==2020.1
PyYAML==6.0
rdflib==6.2.0
redis==4.4.0rc2
requests==2.28.1
rich==14.0.0
rich-toolkit==0.14.1
Rx==3.2.0
shellingham==1.5.4
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
typer==0.15.2
typing_extensions==4.13.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
validators==0.20.0
watchfiles==1.0.5
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
